Physics: Improve object physics code
